Ecto postgres update parsing
