+++ lib/new_relic/transaction/monitor.ex
@@ -95,7 +95,7 @@ defmodule NewRelic.Transaction.Monitor do
          {Ecto.Repo.Supervisor, :start_link, [_repo, otp_app, _adapter, _opts]}, _timestamp},
         state
       ) do
-    NewRelic.Telemetry.EctoSupervisor.start_child(otp_app)
+    NewRelic.Telemetry.Ecto.Supervisor.start_child(otp_app)
     {:noreply, state}
   end
